fixed coordinate issue

+++ include/tact/Cursor.h
@@ -17,13 +17,12 @@ private:
 public:
     Cursor();
     Cursor(int x, int y);
-
     int get_tile_x() { return this->x/32; }
     int get_tile_y() { return this->y/32; }
     Coordinate get_tile_coordinate();
     sf::Sprite get_sprite() { return this->sprite; }
     virtual void moveSprite(float x, float y);
-    void jump_to(float, float);
+    void jump_to(int, int);
 
     friend std::ostream& operator<< (std::ostream& out, const Cursor& cursor);
 };
